[MNG-6118] Applied review comment; Moved Project Collection related files to a new project.collector package
